refactor(@expo/router-server): migrate typed routes logic from `expo-router` to `@expo/router-server`
