+++ packages/@expo/router-server/package.json
@@ -56,7 +56,8 @@
     }
   },
   "devDependencies": {
-    "react-server-dom-webpack": "~19.0.0"
+    "react-server-dom-webpack": "~19.0.0",
+    "tsd": "^0.28.1"
   },
   "dependencies": {
     "debug": "^4.3.4"
