+++ packages/@expo/router-server/src/typed-routes/__tests__/default-types.tsd.ts
