+++ packages/@expo/router-server/src/typed-routes/__tests__/fixtures/default.d.ts
