+++ packages/expo-router/package.json
@@ -69,7 +69,6 @@
     "lint": "expo-module lint",
     "test": "expo-module test",
     "test:rsc": "jest --config jest-rsc.config.js",
-    "test:tsd": "EXPORT_ROUTER_JEST_TSD=true expo-module test",
     "prepublishOnly": "expo-module prepublishOnly",
     "expo-module": "expo-module"
   },
@@ -126,8 +125,7 @@
     "@types/react-dom": "^19.1.7",
     "@types/semver": "^7.7.0",
     "immer": "^10.1.1",
-    "react-server-dom-webpack": "~19.0.0",
-    "tsd": "^0.28.1"
+    "react-server-dom-webpack": "~19.0.0"
   },
   "dependencies": {
     "@expo/metro-runtime": "^6.1.2",
